# infra/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "高级 RAG 知识库独立复现",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/documents/{doc_id}/file": {
      "put": {
        "summary": "Replace File",
        "operationId": "replace_file_api_documents__doc_id__file_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_replace_file_api_documents__doc_id__file_put"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/strategies": {
      "get": {
        "summary": "Strategies",
        "operationId": "strategies_api_strategies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/applications": {
      "get": {
        "summary": "Applications",
        "operationId": "applications_api_applications_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kb_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Application",
        "operationId": "create_application_api_applications_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/{application_id}": {
      "put": {
        "summary": "Update Application",
        "operationId": "update_application_api_applications__application_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Application Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/feedback": {
      "put": {
        "summary": "Submit Feedback",
        "operationId": "submit_feedback_api_runs__run_id__feedback_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feedback": {
      "get": {
        "summary": "Feedback List",
        "operationId": "feedback_list_api_feedback_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kb_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/datasets": {
      "get": {
        "summary": "Datasets",
        "operationId": "datasets_api_datasets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kb_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Dataset",
        "operationId": "create_dataset_api_datasets_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/datasets/generate": {
      "post": {
        "summary": "Generate Dataset",
        "operationId": "generate_dataset_api_datasets_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetGenerate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/datasets/from-feedback": {
      "post": {
        "summary": "Feedback Dataset",
        "operationId": "feedback_dataset_api_datasets_from_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackDatasetInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/evaluations/compare": {
      "post": {
        "summary": "Compare",
        "operationId": "compare_api_evaluations_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparisonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/models": {
      "get": {
        "summary": "Model Status",
        "operationId": "model_status_api_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/models/test": {
      "post": {
        "summary": "Model Test",
        "operationId": "model_test_api_models_test_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ApplicationInput": {
        "properties": {
          "kb_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 1,
            "title": "Kb Id"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "title": "Description",
            "default": ""
          },
          "strategy": {
            "type": "string",
            "enum": [
              "dense",
              "hybrid",
              "full",
              "custom"
            ],
            "title": "Strategy",
            "default": "hybrid"
          },
          "mode": {
            "type": "string",
            "enum": [
              "rag",
              "agent"
            ],
            "title": "Mode",
            "default": "rag"
          },
          "welcome": {
            "type": "string",
            "maxLength": 2000,
            "title": "Welcome",
            "default": "你好，请提出与资料库有关的问题。"
          },
          "fallback": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Fallback",
            "default": "资料库中暂无足够依据回答这个问题。"
          },
          "prompt": {
            "type": "string",
            "maxLength": 8000,
            "title": "Prompt",
            "default": "用中文回答，先给结论，再列依据。"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kb_id",
          "name"
        ],
        "title": "ApplicationInput"
      },
      "Body_replace_file_api_documents__doc_id__file_put": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_replace_file_api_documents__doc_id__file_put"
      },
...
      "ComparisonRequest": {
        "properties": {
          "kb_id": {
            "type": "string",
            "title": "Kb Id"
          },
          "cases": {
            "items": {
              "$ref": "#/components/schemas/EvaluationCase"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Cases"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kb_id",
          "cases"
        ],
        "title": "ComparisonRequest"
      },
...
      "DatasetGenerate": {
        "properties": {
          "kb_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 1,
            "title": "Kb Id"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "default": "文档自动生成测评集"
          },
          "count": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Count",
            "default": 5
          },
          "document_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Document Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kb_id"
        ],
        "title": "DatasetGenerate"
      },
      "DatasetInput": {
        "properties": {
          "kb_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 1,
            "title": "Kb Id"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "cases": {
            "items": {
              "$ref": "#/components/schemas/EvaluationCase"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Cases"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kb_id",
          "name",
          "cases"
        ],
        "title": "DatasetInput"
      },
      "EvaluationCase": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question"
          },
          "reference_answer": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Reference Answer"
          },
          "reference_facts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Reference Facts"
          },
          "relevant_document_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Relevant Document Ids"
          },
          "provenance": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Provenance"
          },
          "relevant_child_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevant Child Ids"
          },
          "annotation_method": {
            "anyOf": [
              {
                "type": "string",
                "const": "human"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annotation Method"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question",
          "reference_answer",
          "reference_facts"
        ],
        "title": "EvaluationCase"
      },
...
      "FeedbackDatasetInput": {
        "properties": {
          "run_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 1,
            "title": "Run Id"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "default": "反馈人工订正测评集"
          },
          "reference_answer": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Reference Answer"
          },
          "reference_facts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Reference Facts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "reference_answer",
          "reference_facts"
        ],
        "title": "FeedbackDatasetInput"
      },
      "FeedbackInput": {
        "properties": {
          "helpful": {
            "type": "boolean",
            "title": "Helpful"
          },
          "comment": {
            "type": "string",
            "maxLength": 2000,
            "title": "Comment",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "helpful"
        ],
        "title": "FeedbackInput"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}